% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{compilation}
\alias{compilation}
\title{Compile a TreePPL model and create inference machinery}
\usage{
compilation(path, args_str)
}
\arguments{
\item{path}{\link[base:character]{base::character} to a treppl model}

\item{args_str}{\link[base:character]{base::character} of options for treeppl compiler}
}
\value{
The path for the compiled TreePPL program.
}
\description{
\code{compilation} compile a TreePPL model and create inference machinery to be
used by \link{tp_run}.
}
