% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{compiled_model_Template-class}
\alias{compiled_model_Template-class}
\alias{compiled_model_Template}
\title{TreePPL model template}
\description{
\code{tp_modelT} template for TreePPL code carrying all the informations necessary
for compiling and running this model efficently
}

