% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{cor_export_num}
\alias{cor_export_num}
\title{Specific correctif for hostrep to be able to export float value}
\usage{
cor_export_num(JSON_str)
}
\description{
Specific correctif for hostrep to be able to export float value
}
