% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_traitment.R
\name{.peel_tree}
\alias{.peel_tree}
\title{Recursive function to go deep in the tree}
\usage{
.peel_tree(
  subtree,
  index,
  pindex,
  lweight,
  lnorm_const,
  mu,
  beta,
  lambda,
  prevAge,
  startState,
  result
)
}
\description{
Recursive function to go deep in the tree
}
