% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{options_to_string}
\alias{options_to_string}
\title{Convert options to a proper string of flags, e.g., \verb{_} to \code{-},
adding \verb{--} in the beginning, spaces between things, etc.}
\usage{
options_to_string(options)
}
\description{
Convert options to a proper string of flags, e.g., \verb{_} to \code{-},
adding \verb{--} in the beginning, spaces between things, etc.
}
