% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_treeppl_output.R
\name{read_treeppl_output}
\alias{read_treeppl_output}
\title{Read output from inference with TreePPL}
\usage{
read_treeppl_output(filename = NULL)
}
\arguments{
\item{filename}{Path to json file with the output from inference with TreePPL}
}
\value{
A data frame with one row for each sample and columns for parameters and weights for each run.
If the file contains the output from multiple runs, returns a list of data frames.
}
\description{
Read output from inference with TreePPL
}
