% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_treeppl.R
\name{run_treeppl}
\alias{run_treeppl}
\title{Compile and run a TreePPL program}
\usage{
run_treeppl(
  dir = NULL,
  source = NULL,
  method = "smc-bpf",
  data_path = NULL,
  samples = 1000
)
}
\arguments{
\item{dir}{The directory that contains the TreePPL code and input file. The executable and the output files will be written here.}

\item{source}{Name of file with TreePPL program.}

\item{method}{Inference method to be used.}

\item{data_path}{Path to JSON file with input data.}

\item{samples}{The number of samples during inference.}
}
\value{
A data frame with sampled values, log weights, normalized weights, and the normalizing constant for all samples.
}
\description{
Compile and run a TreePPL program
}
\examples{
\dontrun{
  coinflips <- tibble(coinflips = sample(c(TRUE, FALSE), 20, replace = TRUE))
  output <- run_treeppl(dir = system.file("extdata", package = "treepplr"),
                        source = "coin.tppl", data = coinflips, samples = 10)
  if(rlang::is_installed("ggplot2")) {
    ggplot2::ggplot(output) +
    ggplot2::geom_col(ggplot2::aes(.data$samples, .data$nweights), width = 0.005) +
    ggplot2::theme_bw()
  }
}
}
