% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{sampler_T-class}
\alias{sampler_T-class}
\alias{sampler_T}
\title{TreePPL model template}
\description{
\code{sampler_T} template for TreePPL code carrying all the informations necessary
for compiling and running this model efficently
}

