% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{tp_check}
\alias{tp_check}
\title{Check input for inference with TreePPL}
\usage{
tp_check(model_name, model_string, data_treeppl_input)
}
\arguments{
\item{model_name}{A specific model name call for a specific checker (\code{model_name::basic} call \code{tp_check_basic()}).}

\item{model_string}{A string contain the TreePPL model.}

\item{data_treeppl_input}{A list of R objects contain the TreePPL data.}
}
\value{
tuple : TreePPL model (string) and TreePPL data (r_object)
}
\description{
This function check the conformity of model/data for \code{tp_run()}.
}
