% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{tp_check_input}
\alias{tp_check_input}
\title{Check input for inference with TreePPL}
\usage{
tp_check_input(model, data)
}
\arguments{
\item{model}{A TreePPL model (S3)}

\item{data}{A list with each item name accordingly to the data structure in the
model.}
}
\value{
...
}
\description{
This function checks the input for \code{tp_go()}.
}
