% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tp_compile}
\alias{tp_compile}
\title{Create a TreePPL model}
\usage{
tp_compile(model, method = "mcmc", ...)
}
\arguments{
\item{model}{One of tree options:
\itemize{
\item The full path of the model file that contains the TreePPL code, OR
\item A string with the name of a model supported by treepplr
(see \code{\link[=tp_model_library]{tp_model_library()}}), OR
\item A string containing the entire TreePPL code.
}}

\item{method}{Inference method to be used. See \code{\link[=tp_compile_methods]{tp_compile_methods()}}
for all supported methods.}

\item{...}{See \code{\link[=tp_compile_options]{tp_compile_options()}} for all supported arguments.}
}
\value{
sampler from a sampler_T
}
\description{
\code{tp_compile} takes TreePPL model and create a sampler to be used by
\code{\link[=tp_run]{tp_run()}}.
}
