% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tp_compile_methods}
\alias{tp_compile_methods}
\title{Methods to \code{\link[=tp_compile]{tp_compile()}}}
\usage{
tp_compile_methods()
}
\value{
A data frame with methods that can be passed to \code{\link[=tp_compile]{tp_compile()}}
}
\description{
Methods to \code{\link[=tp_compile]{tp_compile()}}
}
