% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tp_compile_options}
\alias{tp_compile_options}
\title{Options to \code{\link[=tp_compile]{tp_compile()}}}
\usage{
tp_compile_options()
}
\value{
A data frame with options that can be passed to \code{\link[=tp_compile]{tp_compile()}}
}
\description{
Options to \code{\link[=tp_compile]{tp_compile()}}
}
