% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_data_stored}
\alias{tp_data_stored}
\title{Provide data filenames stored by user in \code{\link[base:tempfile]{base::tempdir()}}) using
\code{\link[=tp_write]{tp_write()}})}
\usage{
tp_data_stored()
}
\value{
A list of data filenames.
}
\description{
Provide a list of all the data filenames currently
stored in \code{\link[base:tempfile]{base::tempdir()}}). They can also be use to verify if model are
attach to these data (see \code{\link[=tp_model_stored]{tp_model_stored()}})
}
