% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tp_expected_input}
\alias{tp_expected_input}
\title{List expected input variables for a model}
\usage{
tp_expected_input(model)
}
\arguments{
\item{model}{A TreePPL model}
}
\value{
The expected input data for a given TreePPL model.
}
\description{
List expected input variables for a model
}
