% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_go}
\alias{tp_go}
\title{Compile and run a TreePPL program}
\usage{
tp_go(
  dir_path = NULL,
  project_name = "input",
  src_name = "out",
  method = "smc-bpf",
  samples = 1000
)
}
\arguments{
\item{dir_path}{a character vector giving the directory name.}

\item{project_name}{a character vector giving the project name.}

\item{src_name}{a character vector giving the output name.}

\item{method}{a character vector giving the inference method name.}

\item{samples}{a \link[base:integer]{base::integer} giving the number of samples.}
}
\value{
Path to the tmp R directory where output file was written.
}
\description{
\code{tp_go} take paths, execute TreePPL and return current \link[base:tempfile]{base::tempdir} path.
}
\details{
This function takes TreePPL code (.tppl) and data (.json) files, compile TreePPL code using \code{\link[=tp_compile]{tp_compile()}}
and writing TreePPL output using \code{\link[=tp_run]{tp_run()}}.

TreePPL need to be install on your computer and the PATH set for R/RSTUDIO (see \href{https://www.google.com}{install} manual).
The executable and the output files will be written in R's \code{\link[base:tempfile]{base::tempdir()}}.

\code{dir_path} : The directory that contains the TreePPL code and data file. If NULL, use \code{\link[=tp_tempdir]{tp_tempdir()}} output.

\code{project_name} : Name of the TreePPL code and data file in dir_path (without extension).

\code{src_name} : File name for TreePPL program and TreePPL output.

\code{method} : Inference method to be used.

\code{samples} : The number of samples during inference.
}
