% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{tp_import_data}
\alias{tp_import_data}
\title{Import model for TreePPL program}
\usage{
tp_import_data(dir_path, data_name)
}
\arguments{
\item{dir_path}{a character vector giving the directory name.}

\item{data_name}{a character vector giving the data file name.}
}
\value{
Structured list of TreePPL data
}
\description{
\code{tp_import_data} takes a directory path and out an data to be check by \code{tp_check()}.
}
\details{
This function takes TreePPL data (.json) files and import them.

\code{dir_path} : The directory that contains the TreePPL data If NULL, use \code{\link[=tp_tempdir]{tp_tempdir()}} output.

\code{data_name} : Name of the TreePPL data in dir_path (without extension).
}
