% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{tp_import_model}
\alias{tp_import_model}
\title{Import model for TreePPL program}
\usage{
tp_import_model(dir_path, model_name)
}
\arguments{
\item{dir_path}{a character vector giving the directory name.}

\item{model_name}{a character vector giving the model name.}
}
\value{
A string of TreePPL model
}
\description{
\code{tp_import_model} takes a directory path and out an model to be check by \code{tp_check()}.
}
\details{
This function takes TreePPL code (.tppl) files and import them.

\code{dir_path} : The directory that contains the TreePPL code. If NULL, use \code{\link[=tp_tempdir]{tp_tempdir()}} output.

\code{model_name} : Name of the TreePPL code in dir_path (without extension).
}
