% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{tp_input}
\alias{tp_input}
\title{Prepare input for inference with TreePPL}
\usage{
tp_input(model_string, data_treeppl_input)
}
\arguments{
\item{model_string}{A string contain the TreePPL model to be included in the trppl input file.}

\item{data_treeppl_input}{A list of R objects contain the TreePPL data to be included in the TreePPL JSON input file.}
}
\value{
tuple : path to TreePPL file / path to JSON file
}
\description{
This function takes R objects and writes out an JSON file to be used by \code{tp_run()}.
}
