% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_installing_treeppl}
\alias{tp_installing_treeppl}
\title{Platform-dependent treeppl self-contained installation}
\usage{
tp_installing_treeppl(download = TRUE)
}
\arguments{
\item{download}{Will download the associate tpplc version in the dir next
to your local treepplr installation if not present.}
}
\value{
The path for TreePPL compiler.
}
\description{
\code{tp_installing_treeppl} will search for the local version tpplc associate
with the package. Will download it if it's not detected on the computer.
}
