% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{tp_json_list}
\alias{tp_json_list}
\title{Convert a json object to json list}
\usage{
tp_json_list(json)
}
\arguments{
\item{json}{an object of class "json".}
}
\value{
A structured list how contain a json
}
\description{
\code{tp_json} takes an object of class "json" and return a json list
ready to be export as a JSON
}
