% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_list}
\alias{tp_list}
\title{Create a flat list}
\usage{
tp_list(...)
}
\arguments{
\item{...}{Variadic arguments (see details).}
}
\value{
A list.
}
\description{
\code{tp_list} takes a variable number of arguments and returns a list.
}
\details{
This function takes a variable number of arguments, so that users can pass as
arguments either independent lists, or a single structured
list of list (name_arg = value_arg).
}
