% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_treatment.R
\name{tp_map_tree}
\alias{tp_map_tree}
\title{Find the Maximum A Posteriori (MAP) Tree from weighted samples}
\usage{
tp_map_tree(trees_out)
}
\arguments{
\item{trees_out}{The list returned by \link{tp_json_to_phylo}
(containing $trees and $weights)}
}
\value{
The MAP tree as a phylo object
}
\description{
Find the Maximum A Posteriori (MAP) Tree from weighted samples
}
