% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{tp_model}
\alias{tp_model}
\title{Import model for TreePPL program}
\usage{
tp_model(model)
}
\arguments{
\item{model}{polymorphic parameter (see below).}
}
\value{
A TreePPL model (S3)
}
\description{
\code{tp_model} takes TreePPL model code and out an model to be check by \code{\link[=tp_check]{tp_check()}}.
}
\details{
This function takes TreePPL model code and import them.

\code{model} : The full path of the model file that contains the TreePPL model code OR
\code{model} : The model name of one model supported by treepplr package (see \code{\link[=tp_model_name]{tp_model_name()}} for model supported) OR
\code{model} : The full string contains the TreePPL model code.
}
