% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_model_library}
\alias{tp_model_library}
\title{Model names supported by treepplr}
\usage{
tp_model_library()
}
\value{
A list of model names.
}
\description{
Provides a list of all models in the TreePPL model library.
}
