% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_model_name}
\alias{tp_model_name}
\title{Provide normalized model names supported by treepplr}
\usage{
tp_model_name()
}
\value{
A list of model names.
}
\description{
Provide a list of all the model names supported by treepplr.
They can also be use to find data attach to these models
(see \code{\link[=tp_data]{tp_data()}})
}
