% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_model_names}
\alias{tp_model_names}
\title{Model names supported by treepplr}
\usage{
tp_model_names()
}
\value{
A list of model names.
}
\description{
Provides a list of all model names supported by treepplr.
The names can also be used to find data for these models
(see \link{tp_data}).
}
