% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_model_stored}
\alias{tp_model_stored}
\title{Provide model filenames stored by user in \code{\link[base:tempfile]{base::tempdir()}}) using
\code{\link[=tp_write]{tp_write()}})}
\usage{
tp_model_stored()
}
\value{
A list of model filenames.
}
\description{
Provide a list of all the model filenames currently
stored in \code{\link[base:tempfile]{base::tempdir()}}). They can also be use to verify if data are
attach to these models (see \code{\link[=tp_data_stored]{tp_data_stored()}})
}
