% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_parse}
\alias{tp_parse}
\title{Parse a TreePPL output}
\usage{
tp_parse(dir_path = NULL, src_name = NULL)
}
\arguments{
\item{dir_path}{The directory that contains the TreePPL code and input file. The executable and the output files will be written here.}

\item{src_name}{Name of file with TreePPL program.}
}
\value{
A data frame with sampled values, log weights, normalized weights, and the normalizing constant for all samples.
}
\description{
Parse a TreePPL output
}
\examples{
\dontrun{
  coinflips <- tibble(coinflips = sample(c(TRUE, FALSE), 20, replace = TRUE))
  output <- tp_run(dir = system.file("extdata", package = "treepplr"),
                        src_name = "coin.tppl", data = coinflips, samples = 10)
  if(rlang::is_installed("ggplot2")) {
    ggplot2::ggplot(output) +
    ggplot2::geom_col(ggplot2::aes(.data$samples, .data$nweights), width = 0.005) +
    ggplot2::theme_bw()
  }
}
}
