% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_treatment.R
\name{tp_parse_coin}
\alias{tp_parse_coin}
\title{Parse TreePPL json coin output}
\usage{
tp_parse_coin(treeppl_out, n_runs = 1)
}
\arguments{
\item{treeppl_out}{a character vector giving the TreePPL json output.}

\item{n_runs}{a \link[base:integer]{base::integer} giving the number of run (mcmc)/sweap (smc).}
}
\value{
List (n=n_runs) RevBayes dataframe format.
}
\description{
\code{tp_parse_coin} take TreePPL json output and return a data.frame
}
\details{
This function takes a TreePPL json output and write a revBeyes data.fram
format.

\code{treeppl_out} : A TreePPL json output coming from \link{tp_treeppl}.

\code{n_runs} : The number of run (mcmc) / sweap (smc) used for the inference.
}
