% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_treatment.R
\name{tp_parse_host_rep}
\alias{tp_parse_host_rep}
\title{Parse TreePPL json output for host repertoire model}
\usage{
tp_parse_host_rep(treeppl_out, n_runs = 1)
}
\arguments{
\item{treeppl_out}{a character vector giving the TreePPL json output
produced by \link{tp_treeppl}.}

\item{n_runs}{a \link[base:integer]{base::integer} giving the number of runs (MCMC) or sweeps (SMC).}
}
\value{
A list (n = n_runs) of data frames with the output from inference
in TreePPL under the host repertoire evolution model.
}
\description{
\code{tp_parse_host_rep} takes TreePPL json output from inference with the
model of host repertoire evolution and returns a data.frame
}
