% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyjson.R
\name{tp_phyjson}
\alias{tp_phyjson}
\title{Create a phyjson object}
\usage{
tp_phyjson(...)
}
\arguments{
\item{...}{Variadic arguments (see details).}
}
\value{
A phyjson object (S3).
}
\description{
\code{tp_phyjson} takes a variable number of argument and return a phyjson object.
}
\details{
This function takes a variable number of argument, so that users can pass as
arguments either independent lists for each parameter or a single structured
list of list. It's use like a list (name_arg = value_arg, etc).
}
