% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter.R
\name{tp_phyjson_list}
\alias{tp_phyjson_list}
\title{Convert a phyjson_tree object to phyjson_tree list}
\usage{
tp_phyjson_list(phyjson_tree)
}
\arguments{
\item{phyjson_tree}{an object of class "phyjson_tree".}
}
\value{
A structured list how contain a phyjson_tree
}
\description{
\code{tp_phyjson_list} takes an object of class "phyjson_tree" and return a phyjson_tree list ready to be export as a JSON
}
