% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{tp_phylo_2_json}
\alias{tp_phylo_2_json}
\title{Create a from a phylo object}
\usage{
tp_phylo_2_json(phylo_tree)
}
\arguments{
\item{phylo_tree}{an object of class \link[ape:read.tree]{ape::phylo}.}
}
\value{
A json object (S3)
}
\description{
\code{tp_phylo_2_json} takes an object of class "phylo" and return a json
object.
}
