% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyjson.R
\name{tp_phylo_2_phyjson}
\alias{tp_phylo_2_phyjson}
\title{Create a from a phylo object}
\usage{
tp_phylo_2_phyjson(phylo_tree)
}
\arguments{
\item{phylo_tree}{an object of class \link[ape:read.tree]{ape::phylo}.}
}
\value{
A phyjson object (S3)
}
\description{
\code{tp_phylo_2_phyjson} takes an object of class "phylo" and return a phyjson
object.
}
