% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{tp_phylo_2_phyjson_tree}
\alias{tp_phylo_2_phyjson_tree}
\title{Create a phyjson_tree from a phylo object}
\usage{
tp_phylo_2_phyjson_tree(phylo_tree)
}
\arguments{
\item{phylo_tree}{an object of class \link[ape:read.tree]{ape::phylo}.}
}
\value{
A phyjson tree object (S3)
}
\description{
\code{tp_phylo_2_phyjson_tree} takes an object of class "phylo" and return a phyjson_tree object.
}
