% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_convertion.R
\name{tp_phylo_2_tppl_tree}
\alias{tp_phylo_2_tppl_tree}
\title{Convert phylo to a tppl_tree}
\usage{
tp_phylo_2_tppl_tree(phylo_tree)
}
\arguments{
\item{phylo_tree}{an object of class \link[ape:read.tree]{ape::phylo}.}
}
\value{
A pair (root index, tppl_tree)
}
\description{
\code{tp_phylo_2_tppl_tree} takes an object of class \code{phylo} and returns
a tppl_tree.
}
