% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_convertion.R
\name{tp_phylo_2_TreePPL}
\alias{tp_phylo_2_TreePPL}
\title{Convert phylo obj to TreePPL tree}
\usage{
tp_phylo_2_TreePPL(phylo_tree, age = "")
}
\arguments{
\item{phylo_tree}{an object of class \link[ape:read.tree]{ape::phylo}.}

\item{age}{a string that determine the way the age of the node are
calculated (default "branch-length").
\itemize{
\item "branch-length" : Root's age = NA, branch-length from the parent branch
\item "top-down" : Root's age = 0.0, cumulative branch-length from root
\item "down-top" : Leaf's age = 0.0, cumulative branch-length from leaf
}}
}
\value{
A TreePPL json str
}
\description{
\code{tp_phylo_2_TreePPL} takes an object of class \code{phylo} and returns
a TreePPL json str ready to print in a data file for TreePPL use.
}
