% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{tp_run_options}
\alias{tp_run_options}
\title{Options that can be passed to a TreePPL program}
\usage{
tp_run_options()
}
\value{
A string with the output from the executable's help
}
\description{
Options that can be passed to a TreePPL program
}
