% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tp_runtime_options}
\alias{tp_runtime_options}
\title{Options to \code{\link[=tp_run]{tp_run()}}}
\usage{
tp_runtime_options()
}
\value{
A data frame with options that can be passed to \code{\link[=tp_run]{tp_run()}}
}
\description{
Options to \code{\link[=tp_run]{tp_run()}}
}
