% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tp_select_model}
\alias{tp_select_model}
\title{Test script for now, will maybe transform in function later.}
\usage{
tp_select_model(model_name)
}
\description{
Test script for now, will maybe transform in function later.
}
