% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_stored_compiled}
\alias{tp_stored_compiled}
\title{List of compiled models in \link[base:tempfile]{base::tempdir}}
\usage{
tp_stored_compiled()
}
\value{
A list of compiled model file names.
}
\description{
Provides a list of all the compiled model file names currently
stored in \link[base:tempfile]{base::tempdir}.
}
