% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_tempdir}
\alias{tp_tempdir}
\title{Temporary directory for running treeppl}
\usage{
tp_tempdir(temp_dir = NULL, sep = NULL, sub = NULL)
}
\arguments{
\item{temp_dir}{NULL, or a path to be used; if NULL, R's \code{\link[base:tempfile]{base::tempdir()}}
is used.}

\item{sep}{Better ignored; non-default values are passed to
\code{\link[base:normalizePath]{base::normalizePath()}}.}

\item{sub}{Extension for defining a sub-directory within the directory
defined by \link[base:tempfile]{base::tempdir}.}
}
\value{
Normalized path with system-dependent terminal separator.
}
\description{
\code{tp_tempdir} returns a normalized path for a temporary directory where the
executables can read and write temporary files.
}
