% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tp_treeppl_json}
\alias{tp_treeppl_json}
\title{Convert a tppl_tree to TreePPL json str}
\usage{
tp_treeppl_json(tree)
}
\arguments{
\item{tree}{an object of class json.}
}
\value{
A TreePPL json str
}
\description{
\code{tp_treeppl_json} takes an tppl_tree create with \link{tp_phylo_to_tppl_tree}
and returns a list ready to be exported with JSON
}
