% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tp_write_data}
\alias{tp_write_data}
\title{Write data to file}
\usage{
tp_write_data(data_list, data_file_name = "tmp_data_file", dir = tp_tempdir())
}
\arguments{
\item{data_list}{A named list of data input}

\item{data_file_name}{An optional name for the file created}

\item{dir}{The directory where you want to save the data file in JSON format.
Default is \code{\link[base:tempfile]{base::tempdir()}}.}
}
\value{
The path to the created file
}
\description{
Write data to file
}
