% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{tp_write_model}
\alias{tp_write_model}
\title{Write out a custom model to tp_tempdir()}
\usage{
tp_write_model(model, model_file_name = "tmp_model_file")
}
\arguments{
\item{model}{A string containing the entire TreePPL code.}

\item{model_file_name}{An optional name to the file created}
}
\value{
The path to the file created
}
\description{
Write out a custom model to tp_tempdir()
}
