% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_convertion.R
\name{tree_age_cumul}
\alias{tree_age_cumul}
\title{Calculate age in a tppl_tree}
\usage{
tree_age_cumul(tree, root_index, age = "branch-length")
}
\arguments{
\item{tree}{an tppl_tree create with \link{tp_phylo_to_tppl_tree}.}

\item{root_index}{the index of the root in the tppl_tree.}

\item{age}{a string that determine the way the age of the node are
calculated (default "branch-length").
\itemize{
\item "root-to-tip" : Root's age = 0.0, cumulative branch-length from root
\item "tip-to-root" : Leaf's age = 0.0, cumulative branch-length from leaf
}}
}
\value{
A tppl_tree
}
\description{
\code{tree_age_cumul} takes an tppl_tree and returns a json with
branch length accumulation as node's age.
}
