% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{treeppl_check}
\alias{treeppl_check}
\title{Check input for inference with TreePPL}
\usage{
treeppl_check(model_name, model_string, object_list)
}
\arguments{
\item{model_name}{A specific model name call for a specific checker (\code{model_name::basic} call \code{treeppl_check_basic()}).}

\item{model_string}{A string contain the TreePPL model.}

\item{object_list}{A list of R objects contain the TreePPL data.}
}
\value{
tuple : TreePPL model (string) and TreePPL data (r_object)
}
\description{
This function check the conformity of model/data for \code{treeppl_run()}.
}
