% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{treeppl_import}
\alias{treeppl_import}
\title{Import input for inference with TreePPL}
\usage{
treeppl_import(file_path)
}
\arguments{
\item{file_path}{The complete path to the directory where the TreePPL model and TreePPL JSON datafile where.}
}
\value{
tuple : TreePPL model (string) and TreePPL data (r_object)
}
\description{
This function takes a directory path and out an model/data to be check by \code{treeppl_check()}.
}
