% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{treeppl_input}
\alias{treeppl_input}
\title{Prepare input for inference with TreePPL}
\usage{
treeppl_input(model_string, object_list)
}
\arguments{
\item{model_string}{A string contain the TreePPL model to be included in the trppl input file.}

\item{object_list}{A list of R objects contain the TreePPL data to be included in the TreePPL JSON input file.}
}
\value{
tuple : path to TreePPL file / path to JSON file
}
\description{
This function takes R objects and writes out an JSON file to be used by \code{treeppl_run()}.
}
