% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{tp_check_input}
\alias{tp_check_input}
\title{Check input for inference with TreePPL}
\usage{
tp_check_input(model, data)
}
\arguments{
\item{model}{A TreePPL model (S3)}

\item{data}{A json object (S3)}
}
\value{
...
}
\description{
This function check the conformity of model/data for \code{tp_go()}.
}
