% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{tp_compile}
\alias{tp_compile}
\title{Compile a TreePPL model and create inference machinery}
\usage{
tp_compile(
  model,
  method,
  iterations = NULL,
  particles = NULL,
  dir = NULL,
  output = NULL,
  ...
)
}
\arguments{
\item{model}{One of tree options:
\itemize{
\item The full path of the model file that contains the TreePPL code, OR
\item A string with the name of a model supported by treepplr
(see \code{\link[=tp_model_library]{tp_model_library()}}), OR
\item A string containing the entire TreePPL code.
}}

\item{method}{Inference method to be used. See tp_compile_options()
for all supported methods.}

\item{iterations}{The number of MCMC iterations to be run.}

\item{particles}{The number of SMC particles to be run.}

\item{dir}{The directory where you want to save the executable. Default is
\code{\link[base:tempfile]{base::tempdir()}}}

\item{output}{Complete path to the compiled TreePPL program that will be
created. Default is dir/\if{html}{\out{<name of the model object>}}.exe}

\item{...}{See tp_compile_options() for all supported arguments.}
}
\value{
The path for the compiled TreePPL program.
}
\description{
\code{tp_compile} compile a TreePPL model and create inference machinery to be
used by \link{tp_run}.
}
