% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_compile}
\alias{tp_compile}
\title{Compile a TreePPL program}
\usage{
tp_compile(
  dir_path = NULL,
  project_name = NULL,
  src_name = NULL,
  method = "smc-bpf"
)
}
\arguments{
\item{dir_path}{The directory that contains the TreePPL code and input file. The executable and the output files will be written here.}

\item{src_name}{Name of file with TreePPL program.}

\item{method}{Inference method to be used.}

\item{data_path}{Path to JSON file with input data.}

\item{samples}{The number of samples during inference.}
}
\value{
A data frame with sampled values, log weights, normalized weights, and the normalizing constant for all samples.
}
\description{
Compile a TreePPL program
}
\examples{
\dontrun{
  coinflips <- tibble(coinflips = sample(c(TRUE, FALSE), 20, replace = TRUE))
  output <- tp_run(dir_path = system.file("extdata", package = "treepplr"),
                        src_name = "coin.tppl", data = coinflips, samples = 10)
  if(rlang::is_installed("ggplot2")) {
    ggplot2::ggplot(output) +
    ggplot2::geom_col(ggplot2::aes(.data$samples, .data$nweights), width = 0.005) +
    ggplot2::theme_bw()
  }
}
}
