% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_compile}
\alias{tp_compile}
\title{Compile the TreePPL program}
\usage{
tp_compile(
  model_file_name = "tmp_model_file",
  seed = NULL,
  method = "smc-bpf",
  align = FALSE,
  cps = "none",
  delay = NULL,
  kernel = NULL,
  mcmc_lw_gprob = NULL,
  pmcmc_particles = NULL,
  prune = FALSE,
  subsample = NULL,
  resample = NULL
)
}
\arguments{
\item{model_file_name}{a string giving a name to the temporary model file.}

\item{seed}{a \link[base:numeric]{base::numeric} to use as a random seed.}

\item{method}{a string giving the inference method name.}

\item{align}{a \link[base:logical]{base::logical}. Whether or not to align the model.}

\item{cps}{a string giving the configuration of CPS transformation.}

\item{delay}{a string giving the configuration of delayed sampling.}

\item{kernel}{a \link[base:numeric]{base::numeric} value giving the driftScale for driftKernel
in MCMC.}

\item{mcmc_lw_gprob}{a \link[base:numeric]{base::numeric} probability of performing a global
MCMC step.}

\item{pmcmc_particles}{a \link[base:integer]{base::integer} number of particles for the SMC
proposal computation.}

\item{prune}{a \link[base:logical]{base::logical}. Whether or not to prune the model.}

\item{subsample}{a \link[base:integer]{base::integer} number of draws to subsample from the
posterior distribution.}

\item{resample}{a string giving the selected resample placement method.}

\item{samples}{a \link[base:integer]{base::integer} giving the number of samples (MCMC) /
particles (SMC).}
}
\value{
The directory whreŕe the compiled file is stored.
}
\description{
\code{tp_compile} compiles a TreePPL model to be used by \link{tp_run}.
}
