% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tp_compile_options}
\alias{tp_compile_options}
\title{Options that can be passed to TreePPL compiler}
\usage{
tp_compile_options()
}
\value{
A data frame with some outputs from the compiler's help <tpplc --help>
}
\description{
Options that can be passed to TreePPL compiler
}
