% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{tp_data}
\alias{tp_data}
\title{Import data for TreePPL program}
\usage{
tp_data(data_input)
}
\arguments{
\item{data_input}{polymorphic parameter (see below).}
}
\value{
A phyjson data (S3), see \code{\link[=tp_phyjson]{tp_phyjson()}} for further
details.
}
\description{
\code{tp_data} takes TreePPL program data and out an data to be use by
\code{\link[=tp_treeppl]{tp_treeppl()}}.
}
\details{
This function takes TreePPL program data and import them.

\code{data_input} : The full path of the data file that contains
the TreePPL program data OR
\code{data_input} : The model name of one model supported by treepplr package
(see \code{\link[=tp_model_name]{tp_model_name()}} for exemple data supported) OR
\code{data_input} : The list (or strucutred list) containing TreePPL program data.
}
