% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tp_data}
\alias{tp_data}
\title{Import data for TreePPL program}
\usage{
tp_data(data_input)
}
\arguments{
\item{data_input}{One of the following options:
\itemize{
\item A list (or structured list) containing TreePPL data, OR
\item The full path of a multiple sequence alignment in fasta (.fasta, .fas)
or nexus (.nexus, .nex) format, OR
\item For test data, a string with the name of a model supported by treepplr
(see \code{\link[=tp_model_library]{tp_model_library()}}).
}}

\item{dir}{The directory where you want to save the data file in JSON format.
Default is \code{\link[base:tempfile]{base::tempdir()}}.}
}
\value{
The path for the data file that will be used by \code{\link[=tp_run]{tp_run()}}.
}
\description{
Prepare data input for \code{\link[=tp_run]{tp_run()}}.
}
\examples{
\donttest{
# Example using a model name supported by TreePPL
input <- tp_data("tree_inference")
input

# Example using an internal FASTA file (same input data as before, but in fasta format)
fasta_file <- system.file("extdata", "tree_inference.fasta", package = "treepplr")
input <- tp_data(fasta_file)
input
}

}
