% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tp_data}
\alias{tp_data}
\title{Import data for TreePPL program}
\usage{
tp_data(data_input)
}
\arguments{
\item{data_input}{One of tree options:
\itemize{
\item A list (or structured list) containing TreePPL data, OR
\item A string with the name of a model supported by treepplr
(see \code{\link[=tp_model_library]{tp_model_library()}}), OR
\item The full path to the file in TreePPL JSON format containing the data.
}}

\item{dir}{The directory where you want to save the data file in JSON format.
Default is \code{\link[base:tempfile]{base::tempdir()}}.}
}
\value{
The path for the data file that will be used by \link{tp_run}.
}
\description{
Prepare data input for \code{\link[=tp_run]{tp_run()}}.
}
