% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{tp_data}
\alias{tp_data}
\title{Import data for TreePPL program}
\usage{
tp_data(data)
}
\arguments{
\item{data}{polymorphic parameter (see below).}
}
\value{
A TreePPL data (S3)
}
\description{
\code{tp_data} takes TreePPL data and out an data to be check by \code{\link[=tp_check]{tp_check()}}.
}
\details{
This function takes TreePPL data and import them.

\code{data} : The full path of the data file that contains the TreePPL data OR
\code{data} : The model name of one model supported by treepplr package (see \code{\link[=tp_model_name]{tp_model_name()}} for exemple data supported) OR
\code{data} : The full string contains the TreePPL data.
}
