% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{tp_json}
\alias{tp_json}
\title{Create a json object}
\usage{
tp_json(...)
}
\arguments{
\item{...}{Variadic arguments (see details).}
}
\value{
A json object (S3).
}
\description{
\code{tp_json} takes a variable number of argument and return a json object.
}
\details{
This function takes a variable number of argument, so that users can pass as
arguments either independent lists for each parameter or a single structured
list of list. It's use like a list (name_arg = value_arg, etc).
}
