% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_convertion.R
\name{tp_json_to_phylo}
\alias{tp_json_to_phylo}
\title{Convert TreePPL multi-line JSON to R phylo/multiPhylo object with associated
weights}
\usage{
tp_json_to_phylo(json_out)
}
\arguments{
\item{json_out}{One of two options:
\itemize{
\item A list of TreePPL output in parsed JSON format (output from
\code{\link[=tp_treeppl]{tp_treeppl()}}), OR
\item The full path of the json file containing the TreePPL output.
}}
}
\value{
A list with two elements:
$trees: A 'phylo' object (if one tree) or 'multiPhylo' object (if multiple).
$weights: A numeric vector of sample weights.
}
\description{
\code{tp_json_to_phylo} takes the path to a TreePPL json output and returns an
object of class \code{phylo}.
}
