% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_treatment.R
\name{tp_mcmc_convergence}
\alias{tp_mcmc_convergence}
\title{Check for convergence across multiple MCMC runs.}
\usage{
tp_mcmc_convergence(treeppl_out)
}
\arguments{
\item{treeppl_out}{a data frame outputted by \code{\link[=tp_parse_mcmc]{tp_parse_mcmc()}}.}
}
\value{
Gelman and Rubin's convergence diagnostic
}
\description{
Check for convergence across multiple MCMC runs.
}
