% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{tp_model}
\alias{tp_model}
\title{Import a TreePPL model}
\usage{
tp_model(model_input)
}
\arguments{
\item{model_input}{One of tree options:
\itemize{
\item The full path of the model file that contains the TreePPL code, OR
\item A string with the name of a model supported by treepplr
(see \code{\link[=tp_model_names]{tp_model_names()}}), OR
\item A string containing the entire TreePPL code.
}}
}
\value{
A TreePPL model (S3). A structured list with the
string containing the TreePPL model and a class
(\code{\link[=tp_model_names]{tp_model_names()}} or "custom")
}
\description{
\code{tp_model} takes TreePPL code and prepares it to be used by
\code{\link[=tp_treeppl]{tp_treeppl()}}.
}
