% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{tp_model}
\alias{tp_model}
\title{Import a TreePPL model}
\usage{
tp_model(model_input)
}
\arguments{
\item{model_input}{One of tree options:
\itemize{
\item The full path of the model file that contains the TreePPL code, OR
\item A string with the name of a model supported by treepplr
(see \code{\link[=tp_model_library]{tp_model_library()}}), OR
\item A string containing the entire TreePPL code.
}}
}
\value{
The path to the TreePPL model file
}
\description{
\code{tp_model} takes TreePPL code and prepares it to be used by
\code{\link[=tp_compile]{tp_compile()}}.
}
