% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_traitment.R
\name{tp_parse}
\alias{tp_parse}
\title{Parse TreePPL json output}
\usage{
tp_parse(treeppl_out, run = 1)
}
\arguments{
\item{treeppl_out}{a character vector giving the TreePPL json output.}

\item{run}{a \link[base:integer]{base::integer} giving the number of run (mcmc) or sweap (smc).}
}
\value{
List (n=run) RevBayes dataframe format.
}
\description{
\code{tp_parse} take TreePPL json output and return a data.frame
}
\details{
This function takes a TreePPL json output and write a revBeyes data.fram format.

\code{treeppl_out} : A TreePPL json output coming from \link{tp_treeppl}.

\code{run} : The number of run (mcmc) / sweap (smc) used for the inference.
}
