% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_parse}
\alias{tp_parse}
\title{Parse TreePPL json output}
\usage{
tp_parse(dir_path = NULL, src_name = "out")
}
\arguments{
\item{dir_path}{a character vector giving the TreePPL json output directory name.}

\item{src_name}{a character vector giving data.frame output name.}
}
\value{
RevBayes data.fram format.
}
\description{
\code{tp_parse} take TreePPL json output and return a data.frame
}
\details{
This function takes a TreePPL json output file compile and write a revBeyes data.fram format.

\code{dir_path} : The directory that contains TreePPL json output coming from \link{tp_run}. If NULL, use \code{\link[=tp_tempdir]{tp_tempdir()}} output.

\code{project_name} : Name of the TreePPL json output file in dir_path (without extension).
}
