% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_treatment.R
\name{tp_parse}
\alias{tp_parse}
\title{Parse simple TreePPL json output}
\usage{
tp_parse(treeppl_out)
}
\arguments{
\item{treeppl_out}{a character vector giving the TreePPL json output
produced by \link{tp_treeppl}.}
}
\value{
A data frame with the output from inference in TreePPL.
}
\description{
\code{tp_parse} takes TreePPL json output and returns a data.frame
}
