% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_treatment.R
\name{tp_parse_mcmc}
\alias{tp_parse_mcmc}
\title{Parse simple TreePPL json MCMC output}
\usage{
tp_parse_mcmc(treeppl_out)
}
\arguments{
\item{treeppl_out}{a character vector giving the TreePPL json output
produced by \link{tp_run} using an MCMC method.}
}
\value{
A data frame with the output from inference in TreePPL.
}
\description{
\code{tp_parse_mcmc} takes TreePPL json MCMC output and returns a data.frame
}
