% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_treatment.R
\name{tp_parse_smc}
\alias{tp_parse_smc}
\title{Parse simple TreePPL json SMC output}
\usage{
tp_parse_smc(treeppl_out)
}
\arguments{
\item{treeppl_out}{a character vector giving the TreePPL json output
produced by \link{tp_run} using an SMC method.}
}
\value{
A data frame with the output from inference in TreePPL.
}
\description{
\code{tp_parse_smc} takes TreePPL json SMC output and returns a data.frame
}
