% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyjson.R
\name{tp_phyjson_list}
\alias{tp_phyjson_list}
\title{Convert a phyjson object to phyjson list}
\usage{
tp_phyjson_list(phyjson)
}
\arguments{
\item{phyjson}{an object of class "phyjson".}
}
\value{
A structured list how contain a phyjson
}
\description{
\code{tp_phyjson} takes an object of class "phyjson" and return a phyjson list
ready to be export as a JSON
}
