% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{tp_run}
\alias{tp_run}
\title{Run a TreePPL program}
\usage{
tp_run(
  compiled_model,
  data,
  n_runs = NULL,
  n_sweeps = NULL,
  dir = NULL,
  out_file_name = "out",
  ...
)
}
\arguments{
\item{compiled_model}{a \link[base:character]{base::character} with the full path to the compiled model
outputted by \link{tp_compile}.}

\item{data}{a \link[base:character]{base::character} with the full path to the data file in JSON format
outputted by \link{tp_data}.}

\item{n_runs}{When using MCMC, a \link[base:integer]{base::integer} giving the number of runs to be done.}

\item{n_sweeps}{When using SMC, a \link[base:integer]{base::integer} giving the number of SMC sweeps to be done.}

\item{dir}{a \link[base:character]{base::character} with the full path to the directory where you
want to save the output. Default is \code{\link[base:tempfile]{base::tempdir()}}.}

\item{out_file_name}{a \link[base:character]{base::character} with the name of the output file in
JSON format. Default is "out".}

\item{...}{See \link{tp_run_options} for all supported arguments.}
}
\value{
A list of TreePPL output in parsed JSON format.
}
\description{
Run TreePPL and return output.
}
\examples{
# When using SMC
# compile model and create SMC inference machinery
exe_path <- tp_compile(model = "coin", method = "smc-bpf", particles = 2000)

# prepare data
data_path <- tp_data(data_input = "coin")

# run TreePPL
result <- tp_run(exe_path, data_path, n_sweeps = 2)


# When using MCMC
# compile model and create MCMC inference machinery
exe_path <- tp_compile(model = "coin", method = "mcmc-naive", iterations = 2000)

# prepare data
data_path <- tp_data(data_input = "coin")

# run TreePPL
result <- tp_run(exe_path, data_path, n_runs = 2)
}
