% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_run}
\alias{tp_run}
\title{Run a TreePPL program}
\usage{
tp_run(
  model_file_name = "tmp_model_file",
  data_file_name = "tmp_data_file",
  samples = 1000,
  n_runs = "1"
)
}
\arguments{
\item{model_file_name}{a string giving a name to the temporary model file.}

\item{data_file_name}{a string giving a name to the temporary data file.}

\item{samples}{a \link[base:integer]{base::integer} giving the number of samples (MCMC) /
particles (SMC).}

\item{n_runs}{a \link[base:integer]{base::integer} giving the number of runs (MCMC) / sweeps (SMC).}
}
\value{
TreePPL output in JSON format.
}
\description{
\code{tp_treeppl} executes TreePPL and returns the inference output in JSON format.
}
