% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_run}
\alias{tp_run}
\title{Run a TreePPL program}
\usage{
tp_run(
  model_name = "input",
  data_name = "input",
  method = "smc-bpf",
  samples = 1000,
  run = "1"
)
}
\description{
Run a TreePPL program
}
