% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tp_runtime_options}
\alias{tp_runtime_options}
\title{Options that can be passed to TreePPL run}
\usage{
tp_runtime_options()
}
\value{
A data frame with with some outputs from compiler's help <tpplc --help>
}
\description{
Options that can be passed to TreePPL run
}
