% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_stored_data}
\alias{tp_stored_data}
\title{Data file names stored by user in \link[base:tempfile]{base::tempdir} using
\link{tp_write}}
\usage{
tp_stored_data()
}
\value{
A list of data file names.
}
\description{
Provides a list of all the data file names currently
stored in \link[base:tempfile]{base::tempdir}. To verify if there is a
matching model file, see \link{tp_stored_model}.
}
