% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tp_tempdir}
\alias{tp_tempdir}
\title{Provide normalized names and make sure the dir exists}
\usage{
tp_tempdir(temp_dir = NULL, sep = NULL, sub = NULL)
}
\arguments{
\item{temp_dir}{NULL, or a path to be used; if NULL, R's \code{tempdir()} is used.}

\item{sep}{Better ignored; non-default values are passed to \code{normalizePath()}.}

\item{sub}{Extension for defining a sub-directory within the directory defined by \code{temp_dir}}
}
\value{
Normalized path with system-dependent terminal separator.
}
\description{
This provides a temporary directory where the executables can read and write temporaty files. Its path is returned in normalized format \bold{with} system-dependent terminal separator.
}
\examples{
tp_tempdir(NULL)
}
