% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_treeppl}
\alias{tp_treeppl}
\title{Compile and run a TreePPL program}
\usage{
tp_treeppl(
  model = NULL,
  model_name = "input",
  data = NULL,
  data_name = "input",
  compile_model = TRUE,
  method = "smc-bpf",
  samples = 1000,
  subsample = NULL,
  run = 1
)
}
\arguments{
\item{model}{a TreePPL model (S3).}

\item{model_name}{a character vector giving a model name.}

\item{data}{a phyjson object (S3).}

\item{data_name}{a character vector giving a data name.}

\item{compile_model}{a \link[base:logical]{base::logical} telling if the model need to be compile again.}

\item{method}{a character vector giving the inference method name.}

\item{samples}{a \link[base:integer]{base::integer} giving the number of samples (mcmc) or particules (smc).}

\item{subsample}{a \link[base:logical]{base::logical} telling if  ...}

\item{run}{a \link[base:integer]{base::integer} giving the number of run (mcmc) or sweap (smc).}
}
\value{
Path to R directory where output file was written.
}
\description{
\code{tp_treeppl} execute TreePPL and return TreePPL output (string JSON format).
}
\details{
This function takes TreePPL model code and data, compile TreePPL model, run it with data
and returning TreePPL output.

TreePPL need to be install on your computer and the PATH set for R/RSTUDIO (see \href{https://treeppl.org/docs/Howtos}{install} manual).
The executable and the output files will be written in R's \code{\link[base:tempfile]{base::tempdir()}}.

\code{model} A TreePPL model (S3), see \link{tp_model} for further details.

\code{model_name}a character vector giving to \link{tp_treeppl} a model name. Can be use to avoid if you have already provide
\code{model} in a previous call of \link{tp_treeppl} to rewrite it.
Can be also use to avoid to write on alredy existing model.

\code{data} A phyjson object (S3), see \link{tp_data} for further details.

\code{data_name} a character vector giving to \link{tp_treeppl} a data name. Can be use to avoid if you have already provide
\code{data} in a previous call of \link{tp_treeppl} to rewrite it.
Can be also use to avoid to write on alredy existing data.

\code{compile_model} : If you want to avoid to recompile.

\code{method} : Inference method to be used (smc, mcmc, etc).

\code{samples} : The number of samples (mcmc) / particules (smc) during inference.

\code{subsample} ...

\code{run} : The number of run (mcmc) / sweap (smc) used for the inference.
}
