% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_write}
\alias{tp_write}
\title{Prepare input for \link{tp_compile}}
\usage{
tp_write(
  model = NULL,
  model_file_name = "tmp_model_file",
  data = NULL,
  data_file_name = "tmp_data_file"
)
}
\arguments{
\item{model}{a TreePPL model (S3).}

\item{model_file_name}{a string giving a name to the temporary model file.}

\item{data}{a json object with all data needed for inference (S3).}

\item{data_file_name}{a string giving a name to the temporary data file.}
}
\description{
This function takes a TreePPL object (the model) and/or a json object (the data) and writes
them in \link[base:tempfile]{base::tempdir} as a .tppl and a .json file, respectively.
}
