% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter.R
\name{tp_write}
\alias{tp_write}
\title{Prepare input for \code{\link[=tp_go]{tp_go()}}}
\usage{
tp_write(model_data_strlist)
}
\arguments{
\item{model_data_strlist}{a structured list coming from \code{\link[=tp_import]{tp_import()}}.}
}
\value{
None.
}
\description{
\code{tp_write} writes an JSON file to be used by \code{\link[=tp_go]{tp_go()}}.
}
\details{
This function takes a structured list of TreePPL code and data files (more details in \code{\link[=tp_import]{tp_import()}}) and write 2 files from them.
}
