% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{tp_write}
\alias{tp_write}
\title{Prepare input for \code{\link[=tp_treeppl]{tp_treeppl()}}}
\usage{
tp_write(model = NULL, model_name = "input", data = NULL, data_name = "input")
}
\arguments{
\item{model_data_strlist}{a structured list coming from \code{\link[=tp_import]{tp_import()}}.}
}
\description{
\code{tp_write} writes an JSON file to be used by \code{\link[=tp_treeppl]{tp_treeppl()}}.
}
\details{
This function takes a structured list of TreePPL code and data files (more details in \code{\link[=tp_import]{tp_import()}}) and write 2 files from them.
}
