% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeppl_input.R
\name{treeppl_input}
\alias{treeppl_input}
\title{Prepare input for inference with TreePPL}
\usage{
treeppl_input(object_list, file_path)
}
\arguments{
\item{object_list}{A list of R objects to be included in the JSON input file.}

\item{file_path}{The complete path to where the JSON file is to be written.}
}
\value{
Path to a JSON file
}
\description{
This function takes R objects and writes out an JSON file to be used by \code{run_treeppl()}.
}
