% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treepplr-package.R
\docType{package}
\name{treepplr-package}
\alias{treepplr}
\alias{treepplr-package}
\title{treepplr: R Interface to TreePPL}
\description{
This package in an interface for using TreePPL programs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/treeppl/treepplr}
  \item \url{http://treeppl.org/treepplr/}
  \item Report bugs at \url{https://github.com/treeppl/treepplr/issues}
}

}
\author{
\strong{Maintainer}: Mariana P Braga \email{mpiresbr@gmail.com} (\href{https://orcid.org/0000-0002-1253-2536}{ORCID})

}
\keyword{internal}
